Revert "Separate postprocessing"
This reverts commit 5368d5c10bbcb21c91a58d8dbf632c4f47d40d7c.
